refactor: Move home-only sections into `home/` folder in components
