+++ src/components/sections/home/Education.tsx
